+++ demo/pages/_app.tsx
@@ -30,6 +30,7 @@ import { HttpTransport } from 'viem'
 import { chainIdToAlchemyNetworkMap } from 'utils/chainIdToAlchemyNetworkMap'
 import { useWalletFilter, WalletFilterProvider } from 'context/walletFilter'
 import { EclipseWalletConnectors } from '@dynamic-labs/eclipse'
+import { AbstractEvmWalletConnectors } from '@dynamic-labs-connectors/abstract-global-wallet-evm'
 import { RelayKitProvider } from '@reservoir0x/relay-kit-ui'
 
 const MoonPayProvider = lazy(() =>
@@ -195,7 +196,8 @@ const AppWrapper: FC<AppWrapperProps> = ({ children }) => {
               SolanaWalletConnectors,
               BitcoinWalletConnectors,
               EclipseWalletConnectors,
-              SuiWalletConnectors
+              SuiWalletConnectors,
+              AbstractEvmWalletConnectors
             ],
             cssOverrides: `
               [data-testid="send-balance-button"] {