+++ packages/ui/src/components/common/TokenSelector/ChainFilterSidebar.tsx
@@ -22,6 +22,7 @@ import { groupChains } from '../../../utils/tokenSelector.js'
 type ChainFilterSidebarProps = {
   options: (RelayChain | { id: undefined; name: string })[]
   value: ChainFilterValue
+  isOpen: boolean
   onSelect: (value: ChainFilterValue) => void
   onAnalyticEvent?: (eventName: string, data?: any) => void
   onInputRef?: (element: HTMLInputElement | null) => void
@@ -39,6 +40,7 @@ const fuseSearchOptions = {
 export const ChainFilterSidebar: FC<ChainFilterSidebarProps> = ({
   options,
   value,
+  isOpen,
   onSelect,
   onAnalyticEvent,
   onInputRef,
@@ -48,6 +50,7 @@ export const ChainFilterSidebar: FC<ChainFilterSidebarProps> = ({
   const [chainSearchInput, setChainSearchInput] = useState('')
   const chainFuse = new Fuse(options, fuseSearchOptions)
   const activeChainRef = useRef<HTMLButtonElement | null>(null)
+  const [hasScrolledOnOpen, setHasScrolledOnOpen] = useState(false)
 
   const { allChainsOption, popularChains, alphabeticalChains } = useMemo(
     () => groupChains(options, popularChainIds),
@@ -71,13 +74,21 @@ export const ChainFilterSidebar: FC<ChainFilterSidebarProps> = ({
   }, [chainSearchInput, chainFuse])
 
   useEffect(() => {
-    if (activeChainRef.current) {
-      activeChainRef.current.scrollIntoView({
-        behavior: 'auto',
-        block: 'nearest'
-      })
+    if (activeChainRef.current && isOpen && !hasScrolledOnOpen) {
+      const timer = setTimeout(() => {
+        activeChainRef.current?.scrollIntoView({
+          behavior: 'instant',
+          block: 'nearest'
+        })
+        setHasScrolledOnOpen(true)
+      }, 100)
+
+      return () => clearTimeout(timer)
+    } else if (!isOpen) {
+      setHasScrolledOnOpen(false)
+      activeChainRef.current = null
     }
-  }, [filteredChains, value])
+  }, [activeChainRef.current, isOpen, hasScrolledOnOpen])
 
   return (
     <Flex
@@ -222,6 +233,7 @@ export const ChainFilterSidebar: FC<ChainFilterSidebarProps> = ({
                             tokenSearchInputRef?.focus()
                           }
                         }}
+                        activeChainRef={activeChainRef}
                         value={chain.id?.toString()}
                         key={chain.id?.toString()}
                       />
@@ -246,6 +258,7 @@ export const ChainFilterSidebar: FC<ChainFilterSidebarProps> = ({
                         tokenSearchInputRef?.focus()
                       }
                     }}
+                    activeChainRef={activeChainRef}
                     value={chain.id?.toString()}
                     key={chain.id?.toString()}
                   />
@@ -265,6 +278,7 @@ type ChainFilterRowProps = {
   onClick?: (e: React.MouseEvent) => void
   tag?: string
   value: string
+  activeChainRef?: React.RefObject<HTMLButtonElement>
 }
 
 const ChainFilterRow: FC<ChainFilterRowProps> = ({
@@ -272,7 +286,8 @@ const ChainFilterRow: FC<ChainFilterRowProps> = ({
   isActive,
   onClick,
   tag,
-  value
+  value,
+  activeChainRef
 }) => {
   return (
     <AccessibleListItem value={value} asChild>
@@ -280,6 +295,7 @@ const ChainFilterRow: FC<ChainFilterRowProps> = ({
         color="ghost"
         size="none"
         onClick={onClick}
+        ref={isActive ? activeChainRef : null}
         css={{
           p: '2',
           display: 'flex',