+++ packages/ui/src/components/widgets/SwapWidgetRenderer.tsx
@@ -711,7 +711,10 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
           (typeof error.message === 'string' &&
             error.message.includes('Plugin Closed')) ||
           (typeof error.message === 'string' &&
-            error.message.includes('denied transaction')))
+            error.message.includes('denied transaction')) ||
+          (typeof error.message === 'string' &&
+            error.message.includes('Failed to initialize request') &&
+            fromChain?.id === 2741)) // Abstract @TODO: remove once privy improves handling rejected transactions
       ) {
         // Close the transaction modal if the user rejects the tx
         setTransactionModalOpen(false)